api/requirements.txt
attrs==22.1.0
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-json-schema==0.0.5
Flask-JWT-Extended==4.4.4
Flask-PyMongo==2.3.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
PyJWT==2.5.0
pymongo==4.2.0
pyrsistent==0.18.1
python-dotenv==0.21.0
Werkzeug==2.2.2
zipp==3.8.1